updated-account-service

+++ account-service/src/main/java/com/account/Dto/BalanceResponse.java
@@ -1,11 +1,13 @@
 package com.account.Dto;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
 public class BalanceResponse {
-	private Double balance;
+	private BigDecimal balance;
 
 }
